refactor(llm-agent): move Anthropic and OpenAI API adapters from llm-agent-server
Move AnthropicApiAdapter and OpenAiApiAdapter implementations and their test
suites from llm-agent-server to llm-agent package. Server-side consumers now
re-export from the new home in @mcp-abap-adt/llm-agent npm package.
