+++ packages/llm-agent/src/api-adapters/__tests__/anthropic-adapter.test.ts
