+++ packages/llm-agent/src/api-adapters/__tests__/openai-adapter.test.ts
